Move class to Dto package.
