+++ src/Api/Dto/Timing.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace AppBundle\Utils;
+namespace AppBundle\Api\Dto;
 
 use Symfony\Component\Serializer\Annotation\Groups;
